Cri mdw (#92)
* cleaning up ejb elements and collitions

* general improvements

* improve in sj cri submodule

* quality pass in adx cri submodule

* axrna and other small module updates

* mwsfdply module

* mvp updates

* sfx and sfh modules

* dolphin card

* softdec module

* matching wmheap function init

* adxt updates

* removing random gaps

* more softdec matching

* SK imporvement

+++ src/dolphin/exi/EXIUart.c
@@ -119,11 +119,12 @@ static int QueueLength(void)
     return 16 - (command >> 24);
 }
 
+/* TODO: [near miss] 97.062500%; donor-supported signed-long transfer length is codegen-neutral; r28/r0 transfer-length lowering remains, stop at clean-C ceiling. */
 int WriteUARTN(void* buffer, unsigned long length)
 {
     unsigned long command;
-    signed long transfer_length;
-    signed long queue_length;
+    int transfer_length;
+    int queue_length;
     char* pointer;
     int locked;
     int error;
